Added code to extract information to database
Added a bunch of variables to hold info which will to inserted into the database.

Code will Insert into two tables: Device and RunLog.  An INSERT into Device will only be made the first time the trojan is run on a device.  An INSERT to RunLog will be made every time the program is run.